Create backend_algo.js
This is a code snippet which we planned to use for the backend logic to analyze the crime log database and show the crime rating(colour graded). But due to time constraints we could not integrate this file with the backend. The logic is explained in the readme file. For testing purpose we have used our own dummy dataset stored in a normal javascript 2d array.